Small modifications to LLM_DQN.

+++ GlucoseLLM/layers/Embed.py
@@ -15,7 +15,7 @@ class TokenEmbedding(nn.Module):
                     m.weight, mode='fan_in', nonlinearity='leaky_relu')
 
     def forward(self, x):
-        x = self.tokenConv(x.permute(0, 2, 1)).transpose(1, 2)
+        x = self.tokenConv(x.permute(0, 2, 1).to(torch.float32)).transpose(1, 2)
         return x
